backend/requirements.dev.txt
Django==5.2.11
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1

psycopg2-binary==2.9.11
dj-database-url==3.1.2
pillow==12.1.0

channels==4.3.2
channels_redis==4.3.0
daphne==4.2.1

django-allauth==65.14.3
dj-rest-auth==7.0.2
django-cors-headers==4.9.0

django-extensions==4.1
requests==2.32.5
redis==7.2.0

django-environ==0.13.0